Making userAccesCode model and login system using fetch
when user send the code to server, by using fetch, server can test if the code user wrote is match with the userAccessCode in the mongoDB. If they are same, user can move to "/notice" and localStorage can get accessCode. Making localStorage have accessCode that user wrote before can persist the access state. Additionally, the userAccessCode is different with managerAccessCode. Code like "if accessCode === userAccessCode" or "if accessCode ===managerAccessCode" can make sure the difference between user view and manager view