Letting website know if user or manager is loggedIn.
Using getSeverSideProps in notice, userAccessCode and managerAccessCode will be send to component. I don't know why but nextjs can send the props as [object Array] like[{balbal:alala, "amama":111}] .So I let the userAccessCode and managerAccessCode [object Array]. So managerAccessCode[0].managerAccessCode is just like a "GM29M". In useEffect, I let browser check the localStorage and know if user is logged in or manager is logged in. Last, I moved the "dbConnect()" into "/" page, not in "/notice".